chore: minor fixes
